Separated schemas and solved small bug
